+++ Input/Input.js
@@ -142,6 +142,15 @@ const InputPopupBase = kind({
 		length: PropTypes.number,
 
 		/**
+		 * Wraps the `InputField`'s value display in a marquee.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
+		marqueeInputField: PropTypes.bool,
+
+		/**
 		 * The maximum length of number input fields.
 		 *
 		 * Overridden by `length` value.
@@ -309,6 +318,7 @@ const InputPopupBase = kind({
 
 	defaultProps: {
 		defaultValue: '',
+		marqueeInputField: false,
 		popupType: 'fullscreen',
 		numberInputField: 'auto',
 		size: 'small',
@@ -376,6 +386,7 @@ const InputPopupBase = kind({
 		disabled,
 		inputAreaClassName,
 		inputFieldSpotlightId,
+		marqueeInputField,
 		noBackButton,
 		noSubmitButton,
 		numberInputField,
@@ -469,6 +480,7 @@ const InputPopupBase = kind({
 									css={css}
 									maxLength={maxLength}
 									minLength={minLength}
+									marqueeContent={marqueeInputField}
 									size={size}
 									autoFocus
 									type={type}