+++ Input/tests/Input-specs.js
@@ -523,4 +523,29 @@ describe('Input specs', () => {
 			expect(buttonSubmit).toBeNull();
 		});
 	});
+
+	describe('marqueeInputField', () => {
+		test('should not render a marquee node in the InputField by default', () => {
+			const {baseElement} = render(
+				<FloatingLayerController>
+					<Input open value="hello" />
+				</FloatingLayerController>
+			);
+			const marqueeText = baseElement.querySelector('.marqueeText');
+
+			expect(marqueeText).toBeNull();
+		});
+
+		test('should render a marquee node in the InputField when marqueeInputField is true', () => {
+			const {baseElement} = render(
+				<FloatingLayerController>
+					<Input open marqueeInputField value="hello" />
+				</FloatingLayerController>
+			);
+			const marqueeText = baseElement.querySelector('.marqueeText');
+
+			expect(marqueeText).not.toBeNull();
+			expect(marqueeText.querySelector('.marquee')).not.toBeNull();
+		});
+	});
 });