+++ samples/sampler/stories/default/Alert.js
@@ -79,6 +79,7 @@ export default {
 
 export const _Alert = (args) => (
 	<Alert
+		buttonDirection={args['buttonDirection']}
 		open={args['open']}
 		onClose={action('onClose')}
 		overlayPosition={args['overlayPosition']}
@@ -97,6 +98,7 @@ export const _Alert = (args) => (
 
 boolean('open', _Alert, Config);
 select('buttons', _Alert, ['no buttons', '1 button', '2 buttons', '3 buttons', '4 buttons'], Config, '2 buttons');
+select('buttonDirection', _Alert, ['auto', 'horizontal', 'vertical'], Config, 'auto');
 text('title', _Alert, Config, 'Fullscreen Alert Title');
 select('type', _Alert, ['fullscreen', 'overlay'], Config);
 select('overlayPosition', _Alert, ['bottom left', 'bottom right', 'center', 'top left', 'top right'], Config);