+++ samples/sampler/stories/default/InputField.js
@@ -23,12 +23,14 @@ export default {
 export const _InputField = (args) => (
 	<InputField
 		autoFocus={args['autoFocus']}
+		caretToEndOnFocus={args['caretToEndOnFocus']}
 		disabled={args['disabled']}
 		dismissOnEnter={args['dismissOnEnter']}
 		iconAfter={args['iconAfter']}
 		iconBefore={args['iconBefore']}
 		invalid={args['invalid']}
 		invalidMessage={args['invalidMessage']}
+		marqueeContent={args['marqueeContent']}
 		onBeforeChange={action('onBeforeChange')}
 		onChange={action('onChange')}
 		placeholder={args['placeholder']}
@@ -38,12 +40,14 @@ export const _InputField = (args) => (
 );
 
 boolean('autoFocus', _InputField, Config);
+boolean('caretToEndOnFocus', _InputField, Config);
 boolean('disabled', _InputField, Config);
 boolean('dismissOnEnter', _InputField, Config);
 select('iconAfter', _InputField, iconNames, Config);
 select('iconBefore', _InputField, iconNames, Config);
 boolean('invalid', _InputField, Config);
 text('invalidMessage', _InputField, Config);
+boolean('marqueeContent', _InputField, Config);
 text('placeholder', _InputField, Config);
 select('size', _InputField, ['small', 'large'], Config);
 select('type', _InputField, prop.type, Config, prop.type[0]);